% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cox_MDTL_ridge}
\alias{plot.cox_MDTL_ridge}
\title{Plot Validation Results for cox_MDTL_ridge Object}
\usage{
\method{plot}{cox_MDTL_ridge}(
  object,
  test_z = NULL,
  test_time = NULL,
  test_delta = NULL,
  test_stratum = NULL,
  criteria = c("loss", "CIndex"),
  ...
)
}
\arguments{
\item{object}{An object of class \code{"cox_MDTL_ridge"}.}

\item{test_z}{Matrix of test covariates.}

\item{test_time}{Vector of test survival times.}

\item{test_delta}{Vector of test status indicators.}

\item{test_stratum}{Vector of test strata.}

\item{criteria}{Metric to plot: \code{"loss"} or \code{"CIndex"}.}

\item{...}{Additional arguments.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots the validation performance against \code{lambda} for MDTL ridge estimates.
}
\examples{
\dontrun{
data(ExampleData_highdim)
train_dat_highdim <- ExampleData_highdim$train
test_dat_highdim <- ExampleData_highdim$test
beta_external_highdim <- ExampleData_highdim$beta_external

mdtl_ridge_est <- cox_MDTL_ridge(z = train_dat_highdim$z,
                                 delta = train_dat_highdim$status,
                                 time = train_dat_highdim$time,
                                 beta = beta_external_highdim,
                                 vcov = NULL,
                                 eta = 0)

plot.cox_MDTL_ridge(mdtl_ridge_est,
                    test_z = test_dat_highdim$z,
                    test_time = test_dat_highdim$time,
                    test_delta = test_dat_highdim$status,
                    criteria = "CIndex")
}
}
