% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_StabSelect.R
\name{plot.StabSelect}
\alias{plot.StabSelect}
\title{Plot Stability Selection Path}
\usage{
\method{plot}{StabSelect}(
  x,
  threshold = 0.75,
  highlight_color = "red",
  background_color = "gray",
  ...
)
}
\arguments{
\item{x}{An object of class \code{StabSelect}.}

\item{threshold}{Numeric. The selection probability threshold (0 to 1). Variables reaching this frequency are highlighted. Default is 0.75.}

\item{highlight_color}{Color for variables that are selected (stable). Default is "red".}

\item{background_color}{Color for variables that are not selected. Default is "gray".}

\item{...}{Additional arguments passed to methods.}
}
\description{
Generates a visualization of the stability paths. Variables that exceed the
specified probability threshold at any point in the path are highlighted.
}
