% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_eval_ncc.R
\name{test_eval_ncc}
\alias{test_eval_ncc}
\title{Evaluate NCC Model: Loss, C-Index, and Brier Score}
\usage{
test_eval_ncc(
  z_ncc,
  case,
  set_id,
  betahat,
  criteria = c("loss", "CIndex", "Brier")
)
}
\arguments{
\item{z_ncc}{Matrix of covariates for NCC data (rows = subjects).}

\item{case}{Integer or logical vector (0/1) indicating cases.}

\item{set_id}{Vector of matched set identifiers.}

\item{betahat}{Numeric vector of estimated coefficients.}

\item{criteria}{"loss", "CIndex", or "Brier".}
}
\value{
Numeric performance metric.
}
\description{
Evaluate NCC Model: Loss, C-Index, and Brier Score
}
\keyword{internal}
