% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_eval_ncc.R
\name{test_eval_ncc}
\alias{test_eval_ncc}
\title{Evaluate NCC Model: Loss and C-Index}
\usage{
test_eval_ncc(z_ncc, case, set_id, betahat, criteria = c("loss", "CIndex"))
}
\arguments{
\item{z_ncc}{Matrix of covariates for NCC data (rows = subjects).}

\item{case}{Integer or logical vector (0/1) indicating cases.}

\item{set_id}{Vector of matched set identifiers.}

\item{betahat}{Numeric vector of estimated coefficients.}

\item{criteria}{"loss" or "CIndex".}
}
\value{
Numeric performance metric.
}
\description{
Evaluate NCC Model: Loss and C-Index
}
\keyword{internal}
